+++ analyser/application/repository_analysis/tests/test_repository_analysis.py
@@ -1,10 +1,9 @@
 from unittest.mock import MagicMock, patch
 
 from application.repository_analysis.repository_analysis import (
-    add_languages_sloc_table,
+    add_file_counts,
     catalogue_repository,
     check_for_excluded_dirs,
-    clone_repo,
 )
 
 FILE_PATH = "application.repository_analysis.repository_analysis"
@@ -13,9 +12,7 @@ FILE_PATH = "application.repository_analysis.repository_analysis"
 @patch(f"{FILE_PATH}.count_files_per_language")
 @patch(f"{FILE_PATH}.catalogue_repository")
 @patch(f"{FILE_PATH}.get_languages")
-@patch(f"{FILE_PATH}.clone_repo")
-def test_add_languages_sloc_table(
-    mock_clone_repo: MagicMock,
+def test_add_file_counts(
     mock_get_languages: MagicMock,
     mock_catalogue_repository: MagicMock,
     mock_count_files_per_language: MagicMock,
@@ -23,40 +20,16 @@ def test_add_languages_sloc_table(
     # Arrange
     markdown_file = MagicMock()
     mock_count_files_per_language.return_value = {"Python": 100, "Javascript": 200}
+    path_to_repo = "application/repos/github-stats"
     # Act
-    add_languages_sloc_table(markdown_file)
+    add_file_counts(markdown_file, "application/repos/github-stats")
     # Assert
     markdown_file.new_table.assert_called_once_with(
         columns=2, rows=3, text=["Language", "File Count", "Python", 100, "Javascript", 200]
     )
-    mock_clone_repo.assert_called_once_with("JackPlowman", "github-stats")
     mock_get_languages.assert_called_once_with()
-    mock_catalogue_repository.assert_called_once_with(mock_clone_repo.return_value, mock_get_languages.return_value)
+    mock_catalogue_repository.assert_called_once_with(path_to_repo, mock_get_languages.return_value)
     mock_count_files_per_language.assert_called_once_with(mock_catalogue_repository.return_value)
-
-
-@patch(f"{FILE_PATH}.Path")
-@patch(f"{FILE_PATH}.Repo")
-def test_clone_repo(mock_repo: MagicMock, mock_path: MagicMock) -> None:
-    # Arrange
-    mock_path.exists.return_value = False
-    # Act
-    clone_repo("JackPlowman", "github-stats")
-    # Assert
-    mock_repo.clone_from.assert_called_once_with(
-        "https://github.com/JackPlowman/github-stats.git", mock_path.return_value
-    )
-
-
-@patch(f"{FILE_PATH}.Path")
-@patch(f"{FILE_PATH}.Repo")
-def test_clone_repo_exists(mock_repo: MagicMock, mock_path: MagicMock) -> None:
-    # Arrange
-    mock_path.exists.return_value = True
-    # Act
-    clone_repo("JackPlowman", "github-stats")
-    # Assert
-    mock_repo.clone_from.assert_not_called()
 
 
 @patch(f"{FILE_PATH}.determine_file_language")
