+++ analyser/application/tests/test_main.py
@@ -1,21 +1,23 @@
 from unittest.mock import MagicMock, patch
 
 from application.main import main
+from application.repository import Repository
 
 FILE_PATH = "application.main"
 
 
-@patch(f"{FILE_PATH}.create_markdown_file")
-@patch(f"{FILE_PATH}.add_languages_sloc_table")
-@patch(f"{FILE_PATH}.set_up_markdown_file")
+@patch(f"{FILE_PATH}.set_up_index_page")
+@patch(f"{FILE_PATH}.analyse_repository")
 def test_main(
-    mock_set_up_markdown_file: MagicMock,
-    mock_add_languages_sloc_table: MagicMock,
-    mock_create_markdown_file: MagicMock,
+    mock_analyse_repository: MagicMock,
+    mock_set_up_index_page: MagicMock,
 ) -> None:
+    # Arrange
+    mock_analyse_repository.return_value = total = 100
     # Act
     main()
     # Assert
-    mock_set_up_markdown_file.assert_called_once_with("index", "GitHub Stats Repository")
-    mock_add_languages_sloc_table.assert_called_once_with(mock_set_up_markdown_file.return_value)
-    mock_create_markdown_file.assert_called_once_with(mock_add_languages_sloc_table.return_value)
+    mock_analyse_repository.assert_called_once_with("JackPlowman/github-stats")
+    mock_set_up_index_page.assert_called_once_with(
+        [Repository("JackPlowman/github-stats", "A repository for analysing GitHub repositories.", total)]
+    )
